+++ scripts/run_cf_plotcutflow.sh
@@ -1,0 +1,6 @@
+law run cf.PlotCutflow --version test --calibrators default --selector default \
+    --datasets mc --processes default --workers 8
+
+# law run cf.PlotCutflow --version test --calibrators default --selector default \
+#     --datasets mc_no_qcd --processes mc_no_qcd --workers 8 \
+#     --yscale log --plot-suffix log